Refatorando status http

+++ src/comic/services/deleteAllComics.service.ts
@@ -1,6 +1,7 @@
 /* eslint-disable prettier/prettier */
-import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
+import { HttpException, HttpStatus, Injectable, NotFoundException } from "@nestjs/common";
 import { PrismaService } from "src/database/PrismaService";
+import { Comic } from "../entities/comic.entity";
 
 
 @Injectable()
@@ -9,11 +10,19 @@ export class deleteAllComicsService{
 
     async deleteAllComics(){
         try {
+
+            const comics = await this.prisma.comic.findMany();
+
+            if(comics.length == 0 ){ throw new Error ('No comic records.');}
+
             await this.prisma.comic.deleteMany({});
-            return { success: true, message: "All comics deleted successfully." };
+
+            return { success: true, message: 'All comics deleted successfully!'};
+
         } catch (error) {
             throw new HttpException(
                 {
+                    sucess: false,
                     message:'Failed to delete comics.',
                     detals:error.message
                 },
